rpc and signer are no longer optional
