+++ parity/main.rs
@@ -46,8 +46,6 @@ extern crate hyper; // for price_info.rs
 extern crate json_ipc_server as jsonipc;
 
 extern crate ethcore_ipc_hypervisor as hypervisor;
-
-#[cfg(feature = "rpc")]
 extern crate ethcore_rpc;
 
 #[cfg(feature = "dapps")]
