+++ parity/rpc.rs
@@ -24,12 +24,8 @@ use jsonipc;
 use rpc_apis;
 use std::fmt;
 
-#[cfg(feature = "rpc")]
 pub use ethcore_rpc::Server as RpcServer;
-#[cfg(feature = "rpc")]
 use ethcore_rpc::{RpcServerError, RpcServer as Server};
-#[cfg(not(feature = "rpc"))]
-pub struct RpcServer;
 
 pub struct HttpConfiguration {
 	pub enabled: bool,
@@ -79,17 +75,6 @@ fn setup_rpc_server(apis: Vec<&str>, deps: &Dependencies) -> Server {
 	rpc_apis::setup_rpc(server, deps.apis.clone(), rpc_apis::ApiSet::List(apis))
 }
 
-#[cfg(not(feature = "rpc"))]
-pub fn setup_http_rpc_server(
-	_deps: &Dependencies,
-	_url: &SocketAddr,
-	_cors_domain: Vec<String>,
-	_apis: Vec<&str>,
-) -> ! {
-	die!("Your Parity version has been compiled without JSON-RPC support.")
-}
-
-#[cfg(feature = "rpc")]
 pub fn setup_http_rpc_server(
 	dependencies: &Dependencies,
 	url: &SocketAddr,
@@ -111,18 +96,12 @@ pub fn setup_http_rpc_server(
 	}
 }
 
-#[cfg(not(feature = "rpc"))]
-pub fn setup_ipc_rpc_server(_dependencies: &Dependencies, _addr: &str, _apis: Vec<&str>) -> ! {
-	die!("Your Parity version has been compiled without JSON-RPC support.")
-}
-
 pub fn new_ipc(conf: IpcConfiguration, deps: &Dependencies) -> Option<jsonipc::Server> {
 	if !conf.enabled { return None; }
 	let apis = conf.apis.split(',').collect();
 	Some(setup_ipc_rpc_server(deps, &conf.socket_addr, apis))
 }
 
-#[cfg(feature = "rpc")]
 pub fn setup_ipc_rpc_server(dependencies: &Dependencies, addr: &str, apis: Vec<&str>) -> jsonipc::Server {
 	let server = setup_rpc_server(apis, dependencies);
 	match server.start_ipc(addr) {
