+++ parity/rpc_apis.rs
@@ -26,13 +26,8 @@ use ethcore::account_provider::AccountProvider;
 use util::network_settings::NetworkSettings;
 use util::network::NetworkService;
 
-#[cfg(feature="rpc")]
 pub use ethcore_rpc::ConfirmationsQueue;
-#[cfg(not(feature="rpc"))]
-#[derive(Default)]
-pub struct ConfirmationsQueue;
 
-#[cfg(feature="rpc")]
 use ethcore_rpc::Extendable;
 
 pub enum Api {
